cast conditionTermination to Int in serial

+++ serial/serial_ant_colony.cpp
@@ -110,7 +110,7 @@ int main(int argc, char* argv[]) {
   loop_counter = 0;
 
   // External loop
-  while (loop_counter < iterations && terminationCondition < ceilf(nAnts * iterations * terminationConditionPercentage)) {
+  while (loop_counter < iterations && terminationCondition < (int) ceilf(nAnts * iterations * terminationConditionPercentage)) {
 
     // printf("Loop nr. : %d\n", loop_counter);
